Rename modules to follow PEP8 naming conventions
